requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
django-cors-headers==4.2.0
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1
gunicorn==20.1.0
phonenumbers==8.13.14
Pillow==9.5.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
whitenoise==6.5.0
